requirements.txt
Django==2.1.3
gunicorn==19.6.0
django-heroku==0.3.1
pymessenger2==3.1.0
dateparser==0.7.0
Pillow==5.3.0